вынесение файлов js в отдельные файлы
