+++ src/js/rendering.js
@@ -1,0 +1,6 @@
+import movieService from '../apiServises/apiService';
+
+// для рендера результатов поиска фильма (необходимо дорабатывать)
+function movieSearchResults() {
+    return movieService.searchMovie()
+}
