+++ src/js/renderingMain.js
@@ -1,0 +1,6 @@
+import movieService from '../apiServises/apiService';
+
+// для рендера главной (необходимо дорабатывать)
+function mainPage() {
+    return movieService.fetchTrending()
+}
